+++ test/runtests.jl
@@ -13,6 +13,11 @@ CSJuMPTestOptimizer() = JuMP.optimizer_with_attributes(CS.Optimizer, "logging" =
 cbc_optimizer = optimizer_with_attributes(Cbc.Optimizer, "logLevel" => 0)
 CSCbcJuMPTestOptimizer() = JuMP.optimizer_with_attributes(CS.Optimizer, "logging" => [], "lp_optimizer" => cbc_optimizer)
 
+macro test_macro_throws(errortype, m)
+    # See https://discourse.julialang.org/t/test-throws-with-macros-after-pr-23533/5878
+    :(@test_throws $(esc(errortype)) try @eval $m catch err; throw(err.error) end)
+end
+
 test_stime = time()
 
 include("general.jl")
@@ -25,6 +30,7 @@ include("options.jl")
 include("moi.jl")
 include("constraints/table.jl")
 include("constraints/indicator.jl")
+include("constraints/reified.jl")
 
 include("lp_solver.jl")